Bot-War/requirements.txt
certifi==2022.12.7
setuptools==65.6.3
pip==22.3.1
attrs==22.1.0
Pillow==9.3.0
multidict==6.0.3
typing-extensions==4.4.0
pymongo==4.3.3
async-timeout==4.0.2
aiohttp==3.8.3
idna==3.4
requests==2.28.1
dnspython==2.2.1
charset-normalizer==2.1.1
discord.py==2.1.0
urllib3==1.26.13
numpy==1.23.5
yarl==1.8.2
frozenlist==1.3.3
aiosignal==1.3.1
wheel==0.38.4
python-dotenv==0.21.0
future==0.18.2
python-bsonjs==0.3.0
auto-py-to-exe==2.32.0
